+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 
 -   `modular-ui` dependency updated
 -   `modular-engine` dependency updated
--   types updated
+-   Types updated
 -   `initApplication` now returns a Promise, to concat it with a `then`
 -   New url query string management (can be disabled inside app.config file/object)